Fixed types

+++ include/types.h
@@ -9,8 +9,8 @@ typedef unsigned short u16;
 typedef short i16;
 typedef unsigned int u32;
 typedef int i32;
-typedef unsigned long u64;
-typedef unsigned long i64;
+typedef unsigned long long u64;
+typedef long long i64;
 typedef size_t size;
 
 typedef u8 bool;
